# docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.InitiatorType:
    enum:
      - salary
      - rent
      - utilities
      - other
      - sale
      - supplier
      - bnpl
    type: string
    x-enum-comments:
      InitiatorTypeBNPL: buy now pay later BNPL transactions
    x-enum-varnames:
      - InitiatorTypeSalary
      - InitiatorTypeRent
      - InitiatorTypeUtilities
      - InitiatorTypeOther
      - InitiatorTypeSales
      - InitiatorTypeSupplier
      - InitiatorTypeBNPL
...
  models.PaymentMethod:
    enum:
      - cash
      - bank
      - terminal
      - online_payment
      - cheque
      - online_transfer
      - undefined
    type: string
    x-enum-varnames:
      - PaymentMethodCash
      - PaymentMethodBank
      - PaymentMethodTerminal
      - OnlineMobileAppPayment
      - Cheque
      - OnlineTransfer
      - PaymentMethodUndefined
...
  models.Transaction:
    properties:
      amount:
        type: integer
      branch_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      payment_method:
        $ref: "#/definitions/models.PaymentMethod"
      type:
        $ref: "#/definitions/models.InitiatorType"
      updated_at:
        type: string
    type: object
...
  models.TransactionOutput:
    properties:
      data:
        items:
          $ref: "#/definitions/models.Transaction"
        type: array
      error:
        items:
          $ref: "#/definitions/models.Error"
        type: array
    type: object
  models.TransactionOutputSingle:
    properties:
      data:
        $ref: "#/definitions/models.Transaction"
      error:
        items:
          $ref: "#/definitions/models.Error"
        type: array
    type: object
  models.TransactionType:
    enum:
      - credit
      - debit
    type: string
    x-enum-comments:
      TransactionTypeCredit:
        credit means - income - when money is gained or received
        into an account
      TransactionTypeDebit:
        debit means - outcome - when money is lost, spent, or
        withdrawn from an account
    x-enum-varnames:
      - TransactionTypeCredit
      - TransactionTypeDebit
...
info:
  contact:
    email: hamidovaslon13@gmail.com
    name: API Support
    url: https://github.com/aslon1213/g4h_pos_erp
  description: This is a ERP/POS API for Magazin.
  title: Magazin ERP/POS API
  version: "1.0"
paths:
...
  /api/transactions/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a transaction from the database by its ID
      parameters:
        - description: Transaction ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: 'message" : "transaction was succesfully deleted'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Delete a transaction by ID
      tags:
        - transactions
    put:
      consumes:
        - application/json
      description: Update transaction details by its ID
      parameters:
        - description: Transaction ID
          in: path
          name: id
          required: true
          type: string
        - description: Transaction amount
          in: query
          name: amount
          type: string
        - description: Transaction description
          in: query
          name: description
          type: string
        - description: Type of transaction
          in: query
          name: type
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: 'message" : "transaction was succesfully updated'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Update a transaction by ID
      tags:
        - transactions
  /api/transactions/{transaction_id}:
    get:
      consumes:
        - application/json
      description: Retrieve a single transaction by its ID
      parameters:
        - description: Transaction ID
          in: path
          name: transaction_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.TransactionOutputSingle"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Get a transaction by ID
      tags:
        - transactions
  /api/transactions/branch/{branch_id}:
    get:
      consumes:
        - application/json
      description: Retrieve transactions based on various query parameters
      parameters:
        - description: Branch ID
          in: path
          name: branch_id
          required: true
          type: string
        - description: Transaction description
          in: query
          name: description
          type: string
        - description: Minimum transaction amount
          in: query
          name: amount_min
          type: integer
        - description: Maximum transaction amount
          in: query
          name: amount_max
          type: integer
        - description: Payment method
          in: query
          name: payment_method
          type: string
        - description: Type of transaction
          in: query
          name: type_of_transaction
          type: string
        - description: Initiator type
          in: query
          name: initiator_type
          type: string
        - description: Minimum date
          in: query
          name: date_min
          type: string
        - description: Maximum date
          in: query
          name: date_max
          type: string
        - description: Page number
          in: query
          name: page
          type: integer
        - description: Number of transactions per page
          in: query
          name: count
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.TransactionOutput"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.Error"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Get transactions by query parameters
      tags:
        - transactions
  /api/transactions/docs/initiator_type:
    get:
      consumes:
        - application/json
      description: Retrieve a list of all possible initiator types for transactions
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/models.InitiatorType"
            type: array
      security:
        - BearerAuth: []
      summary: Get all initiator types
      tags:
        - transactions
  /api/transactions/docs/payment_method:
    get:
      consumes:
        - application/json
      description: Retrieve a list of all possible payment methods
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/models.PaymentMethod"
            type: array
      security:
        - BearerAuth: []
      summary: Get all payment methods
      tags:
        - transactions
  /api/transactions/docs/type:
    get:
      consumes:
        - application/json
      description: Retrieve a list of all possible transaction types
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/models.TransactionType"
            type: array
      security:
        - BearerAuth: []
      summary: Get all transaction types
      tags:
        - transactions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"